Tier 1: Postgres support — kills the SQLite-on-WSL fragility
- db.py: BIGINT for social counts (int4 overflowed on PG); _na() coerces NaN/NaT
  -> NULL and casts numpy scalars to native Python (numpy-2 repr broke psycopg2).
- warehouse/features: BIGINT on count columns.
- tools/migrate_pg.py: SQLite -> Postgres migration + full derived-layer rebuild.
- conftest: honors TEST_DATABASE_URL (run the suite on Postgres); drops scored_posts
  view before tables (PG enforces the dependency).
- test_hardening: backup test is dialect-aware.

Verified on local Postgres 16: migration + full rebuild (fact_mention 1259, all
marts/intelligence/ops), DQ gate 11/11, and the ENTIRE test suite green on BOTH
SQLite and Postgres. Point DATABASE_URL at Postgres/Supabase and everything follows.

+++ analytics/features.py
@@ -24,7 +24,7 @@ DDL = [
         product TEXT PRIMARY KEY, mentions INTEGER, pct_negative REAL, complaints INTEGER,
         avg_score REAL, nps_proxy REAL, updated_at TEXT)""",
     """CREATE TABLE IF NOT EXISTS mart_influencers (
-        author TEXT PRIMARY KEY, author_name TEXT, reach INTEGER, mentions INTEGER,
+        author TEXT PRIMARY KEY, author_name TEXT, reach BIGINT, mentions INTEGER,
         avg_score REAL, stance TEXT, worst_summary TEXT, url TEXT, updated_at TEXT)""",
     """CREATE TABLE IF NOT EXISTS mart_team_queue (
         team TEXT PRIMARY KEY, open_items INTEGER, critical INTEGER, fraud INTEGER,
